♻️ refactor : use-slugged-save-card 분리  #252
